requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
Pillow==10.0.1
python-decouple==3.8
whitenoise==6.6.0
django-crispy-forms==2.1
crispy-tailwind==0.5.0
django-extensions==3.2.3
requests==2.31.0